update llm service routes, remove chat_module and replace with chatbot module
