+++ requirements.txt
@@ -5,6 +5,7 @@ Flask-CORS==4.0.0
 Werkzeug==3.0.3
 python-dotenv==1.0.1
 google-generativeai>=0.3.0
+openai>=1.30.0
 pytest>=7.0.0
 pytest-cov>=4.0.0
 httpx>=0.24.0,<0.28.0
